CA-190961: find_by_file: lookup in stored state
Calling tap-ctl list sometimes seems to be implicated in tap-ctl close taking a
while to be noticed by the tapdisk process.

But we don't need to call tap-ctl list since we now store metadata about tapdisk
processes on this host in /var/run/nonpersistent. This should make the lookup a
lot quicker and relies less on tapdisk co-operation.

This is a bit of a hack for now: we will need to sort out the mapping between
paths and URIs.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>